Final Finishing touch

+++ app/Http/Controllers/ListingController.php
@@ -54,6 +54,10 @@ class ListingController extends Controller
         }
 
     public function update(Request $request, Listing $listing) {
+        //Make sure Logged user is owner
+        if($listing->user_id != auth()->id) {
+            abort(403, 'Unauthorized Action');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -74,7 +78,14 @@ class ListingController extends Controller
 
     //Delete
     public function delete(Listing $listing) {
+        if($listing->user_id != auth()->id) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted successfully');
+    }
+
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
